+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### Version 0.3 (February 5, 2012)
+
+* Using FileCollection across the board.
+* Exposed `srcDirs` as convention property for task `java2htmlConvertCode`.
+* Minor bugfixes.
+
 ### Version 0.2 (January 31, 2012)
 
 * Support for generation of HTML index files - [Issue 8](https://github.com/bmuschko/gradle-java2html-plugin/issues/1).